[Refactor]organismフォルダをorganismsフォルダにrename
